chore:관리자 공지사항 타입 작성
