+++ thiscoding/app/(models)/admin/adminNotices/AdminNoticesListDTO.ts
@@ -1,18 +1,4 @@
-type Role = 'ADMIN' | 'USER' | 'GUEST'; // 예시로 설정, 실제 값에 맞게 수정
-type Social = 'LOCAL' | 'GOOGLE' | 'KAKAO'; // 예시로 설정, 실제 값에 맞게 수정
-
-// User 타입 정의
-interface User {
-  id: number;
-  email: string;
-  password: string;
-  nickname: string;
-  isActivated: boolean;
-  isBanned: boolean;
-  role: Role;
-  social: Social;
-}
-
+import { User } from "@/app/(models)/users/User"
 // Manager 타입 정의 (Manager 엔티티가 어떤 필드를 가지는지 추가로 정의 필요)
 export interface AdminNoticesListDTO {
   id: number;
@@ -20,4 +6,5 @@ export interface AdminNoticesListDTO {
   title: string;
   category: string;
   viewCount: number;
+  createDate: string;
 }
